% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo_module.r
\docType{data}
\name{BazRate}
\alias{BazRate}
\title{as.is' argument equal to TRUE or use the 'colClasses' argument to specify the}
\format{\preformatted{ num 0.85
}}
\usage{
BazRate
}
\description{
as.is' argument equal to TRUE or use the 'colClasses' argument to specify the
}
\keyword{datasets}

