% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo_module.r
\name{buildModule}
\alias{buildModule}
\title{buildModule' function. The second is to invoke the function and assign the
Build the module.}
\usage{
buildModule()
}
\description{
\code{buildModule} builds the module that will carry out the demo models.
}
\details{
This function builds the demo module that simulates foobar behavior. This is done by creating an
environment in which the functions, parameters and specifications are
assigned as described above.
}

