% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo_module.r
\name{demo1}
\alias{demo1}
\title{Demo Function <Replace this with a title for the function.>}
\usage{
demo1()
}
\arguments{
\item{MyArg1}{a numeric value that means nothing <Replace with the name of
the first argument, its type and a description.>}

\item{<Do}{the same thing for every other function argument.>}
}
\value{
A list with two string values. <Describe the tyupe and meaning of the
  return value of the function.
}
\description{
This shows how a function to be included in a module would be defined and
documented <Replace this with a short description of the function.>
}
\details{
This demo function shows the basic structure and formatting of a function and
documentation. This is a function that returns a result that other functions
in the module use and is not saved to the datastore. Note that the return
value is not return in the normal way for functions. Instead it is assigned
to a name in the Module environment. Also note that in the documentation
section that this function is not exported. This means that the function can
not be called directly by package users. It is not necessary for users
because the copy that is included in the module will be called
instead.<Replace this with a detailed description of the function and how it
is used.>
}

