% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo_module.r
\name{demo2}
\alias{demo2}
\title{Demo Function <Replace this with a title for the function.>}
\usage{
demo2()
}
\arguments{
\item{MyArg1}{a numeric value that means nothing <Replace with the name of
the first argument, its type, and a description.>}

\item{<Do}{the same thing for every other function argument.>}
}
\value{
A string vector. <Describe the type and meaning of
  the return value of the function.>
}
\description{
\code{demo2} a demo function <Replace this with a very short description of
the function.>
}
\details{
This demo function shows the basic structure and formatting of a function and
documentation. This is a function that returns a result that is saved to the
datastore.<Replace this with a longer description of the function.>
}

